backend/requirements.txt
agno
python-dotenv
yfinance
